convert domain using the TrainingDataConverter API
